+++ src/config.php
@@ -110,6 +110,16 @@ return [
     // ElementAPI server, you can override what it uses for the `siteUrl` below.
     'siteUrlOverride' => '',
 
+    // The duration of the SEOmatic meta cache in seconds. Null means always cached until explicitly broken
+    // If devMode is on, caches last 30 seconds.
+    'metaCacheDuration' => null,
+
+    // Determines whether the meta container endpoint should be enabled for anonymous frontend access
+    'enableMetaContainerEndpoint' => false,
+
+    // Determines whether the JSON-LD endpoint should be enabled for anonymous frontend access
+    'enableJsonLdEndpoint' => false,
+
     // SeoElementInterface[] The default SeoElement type classes
     'defaultSeoElementTypes' => [
     ],